\name{ModeleADE-package}
\alias{ModeleADE-package}
\alias{ModeleADE}
\docType{package}
\title{
  Outil de Modelisation de l'assurance des emprunteurs
}
\description{
Cet outil permet de modéliser l'assurance des emprunteurs :
Calcul des provisions S1,S2 et IFRS17;
Tarification;
Calcul du Budget et du Business Plan
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
OBEY ISMAEL

Maintainer: OBEY ISMAEL <ismael.obey@natixis.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
