% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Setup}
\alias{Setup}
\title{Enregistre les differentes tables pour utilisation future par les autres fonctions}
\usage{
Setup(TableMortalite, TableEntreeAT, TableMaintienAT, TableMaintienConstitAT)
}
\arguments{
\item{TableMortalite}{Matrice des taux de mortalites . Cette matrice doit avoir 6 colonnes :
Immo_H Perso_H Pro_H Immo_F Perso_F Pro_F .Les ages vont de 18 a 90 ans.}

\item{TableEntreeAT}{Matrice des taux d entree en arret de travail . Cette matrice doit avoir 6 colonnes :
Immo_H Perso_H Pro_H Immo_F Perso_F Pro_F .Les ages vont de 18 a 67 ans.}

\item{TableMaintienAT}{Matrice des taux de maintien en arret de travail.
Il faut une franchise de 3mois sur la table}

\item{TableMaintienConstitAT}{Matrice des capitaux constitutifs en arret de travail.
Il faut une franchise de 3mois sur la table}
}
\description{
Enregistre les differentes tables pour utilisation future par les autres fonctions
}
