% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimulateAgg}
\alias{SimulateAgg}
\title{Title
Projette les modeles points}
\usage{
SimulateAgg(ID_MP, ID_MP_AGG, DUREE_AMORT_MOIS, DUREE_DIFF_MOIS,
  DEBUT_PRET_MOIS, CI_ASSURE, DIFFERE_TYPE, TXNOM_TYPE_MENS, TXNOM_ANN,
  AGE_SOUS_ANN, AGE_FIN_DC_ANN, AGE_FIN_AT_ANN, PART_H, PART_IMMO, PART_PERSO,
  TXACTUMENS, PRIX_PTIA_POURCDCSEUL, ABT_AT_PERTEREV, Periodes_AGG)
}
\arguments{
\item{ID_MP}{Identifiant unique de chaque modele point}

\item{ID_MP_AGG}{Identifiant unique de chaque groupe de modele point}

\item{DUREE_AMORT_MOIS}{Duree amortissement du pret ( en mois )}

\item{DUREE_DIFF_MOIS}{Duree differe du pret ( en mois )}

\item{DEBUT_PRET_MOIS}{Debut du pret à partir de la date de projection  ( en mois )}

\item{CI_ASSURE}{Capital emprunte assure ( quotite comprise )}

\item{DIFFERE_TYPE}{Type de differe ( 1 == Partiel , 2 == Total )}

\item{TXNOM_TYPE_MENS}{Type de mensualisation du taux nominal ( 1 == Proportionnel , 2 == Equivalent )}

\item{TXNOM_ANN}{Taux nominal du pret}

\item{AGE_SOUS_ANN}{Age a la souscription ( en annees )}

\item{AGE_FIN_DC_ANN}{Age fin de la garantie deces ( en annees )}

\item{AGE_FIN_AT_ANN}{Age fin  arret de travail ( en annees )}

\item{PART_H}{Proportion des hommes}

\item{PART_IMMO}{Proportion des prets immos}

\item{PART_PERSO}{Proportion des prets persos}

\item{TXACTUMENS}{Taux actualisation mensuel}

\item{PRIX_PTIA_POURCDCSEUL}{Prix de la PTIA}

\item{ABT_AT_PERTEREV}{Taux abattement pour la perte de revenu}

\item{Periodes_AGG}{Nombre de mois de projection}
}
\value{
Une liste de matrices (Nombre de MP * Nombre de Periodes)
\item{Ech_CRDDebPer}{CRD de debut de periode}
\item{Assu_ProbSurvieCreditDebPer}{Probabilite de survie en debut de periode}
\item{Res_VAN_AssietteCI}{VAN assiette  CI}
\item{Res_VAN_AssietteCRD}{VAN assiette CRD}
\item{Res_VAN_Sin_DCPTIA}{VAN Cumule sinistre DC}
\item{Res_VAN_Sin_ITTIPT}{VAN Cumule sinistre AT}
}
\description{
Title
Projette les modeles points
}
